+++ test/literal.ts
@@ -3,12 +3,20 @@ import {test} from "ava"
 
 test("it should allow only primitives", t => {
     const error = t.throws(() => {
-        const Factory = types.model({
+        types.model({
             complexArg: types.literal({a: 1})
         })
     })
 
-    t.is(error.message, '[mobx-state-tree] Literal types can be built only on top of primitives')
+    t.is(error.message, "[mobx-state-tree] Literal types can be built only on top of primitives")
+})
+
+test("it should fail if not optional and no default provided", (t) => {
+    const Factory = types.literal("hello")
+    const ex = t.throws(() => {
+        Factory.create()
+    })
+    t.deepEqual(ex.message, "[mobx-state-tree] Value \'undefined\' is not assignable to type: hello, expected an instance of hello or a snapshot like \'\"hello\"\' instead.")
 })
 
 test("it should throw if a different type is given", t => {
@@ -17,7 +25,7 @@ test("it should throw if a different type is given", t => {
     })
 
     const error = t.throws(() => {
-        const doc = Factory.create({ shouldBeOne: 2 })
+        Factory.create({ shouldBeOne: 2 })
     })
 
     t.is(error.message, `[mobx-state-tree] Value '{"shouldBeOne":2}' is not assignable to type: TestFactory, expected an instance of TestFactory or a snapshot like \'{ shouldBeOne: 1 }\' instead.`)